Break dummy_app into multiple files
