+++ test/support/dummy_app.ex
@@ -44,50 +44,3 @@ defmodule DummyApp.ErrorView do
 
   def render(_template, _assigns), do: "Error"
 end
-
-defmodule DummyApp.Router do
-  use DummyApp, :router
-
-  pipeline :browser do
-    plug(:accepts, ["html"])
-    plug(:fetch_session)
-    plug(:fetch_live_flash)
-    plug(:put_root_layout, {DummyApp.LayoutView, :root})
-    plug(:protect_from_forgery)
-    plug(:put_secure_browser_headers)
-  end
-
-  pipeline :beacon do
-    plug(BeaconWeb.Plug)
-  end
-
-  scope "/", BeaconWeb do
-    pipe_through(:browser)
-    pipe_through(:beacon)
-
-    live_session :beacon, session: %{"beacon_site" => "my_site"} do
-      live("/*path", PageLive, :path)
-    end
-  end
-end
-
-defmodule DummyApp.Endpoint do
-  # The otp app needs to be beacon otherwise Phoenix LiveView will not be
-  # able to build the static path since it tries to get from `Application.app_dir`
-  # which expects that a real "application" is settled.
-  use Phoenix.Endpoint, otp_app: :beacon
-
-  @session_options [store: :cookie, key: "_dummy_app_key", signing_salt: "secret"]
-
-  socket("/live", Phoenix.LiveView.Socket, websocket: [connect_info: [session: @session_options]])
-
-  plug(Plug.Session, store: :cookie, key: "_app_key", signing_salt: "5Ude+fet")
-
-  plug(DummyApp.Router)
-end
-
-defmodule DummyApp.BeaconDataSource do
-  @behaviour Beacon.DataSource.Behaviour
-
-  def live_data("my_site", ["home"], _params), do: %{vals: ["first", "second", "third"]}
-end
