+++ test/support/dummy_app/beacon_data_source.ex
@@ -1,0 +1,5 @@
+defmodule DummyApp.BeaconDataSource do
+  @behaviour Beacon.DataSource.Behaviour
+
+  def live_data("my_site", ["home"], _params), do: %{vals: ["first", "second", "third"]}
+end
